// src/Mesher/Voxels/Geometry/VoxelMeshVertexStructCursor.d.ts
export declare const VoxelMeshVertexConstants: {
    VertexFloatSize: number;
    VertexByteSize: number;
    PositionOffset: number;
    NormalOffset: number;
    TextureIndexOffset: number;
    UVOffset: number;
    ColorOffset: number;
    VoxelDataOFfset: number;
    MetadataOffset: number;
    DissolutionProximityOffset: number;
    PullStrengthOffset: number;
    SubdivLevelOffset: number;
    PullDirectionBiasOffset: number;
    SubdivAOOffset: number;
    PhNormalizedOffset: number;
};
export declare class VoxelMeshVertexStructCursor {
    static VertexFloatSize: number;
    static VertexByteSize: number;
    static PositionOffset: number;
    static NormalOffset: number;
    static TextureIndexOffset: number;
    static UVOffset: number;
    static ColorOffset: number;
    static VoxelDataOFfset: number;
    static MetadataOffset: number;
    static DissolutionProximityOffset: number;
    static PullStrengthOffset: number;
    static SubdivLevelOffset: number;
    static PullDirectionBiasOffset: number;
    static SubdivAOOffset: number;
    static PhNormalizedOffset: number;
    get positionX(): number;
    set positionX(value: number);
    get positionY(): number;
    set positionY(value: number);
    get positionZ(): number;
    set positionZ(value: number);
    get normalX(): number;
    set normalX(value: number);
    get normalY(): number;
    set normalY(value: number);
    get normalZ(): number;
    set normalZ(value: number);
    get voxelDataX(): number;
    set voxelDataX(value: number);
    get voxelDataY(): number;
    set voxelDataY(value: number);
    get voxelDataZ(): number;
    set voxelDataZ(value: number);
    get voxelDataW(): number;
    set voxelDataW(value: number);
    get metadataX(): number;
    set metadataX(value: number);
    get metadataY(): number;
    set metadataY(value: number);
    get metadataZ(): number;
    set metadataZ(value: number);
    get metadataW(): number;
    set metadataW(value: number);
    get textureIndexX(): number;
    set textureIndexX(value: number);
    get textureIndexY(): number;
    set textureIndexY(value: number);
    get textureIndexZ(): number;
    set textureIndexZ(value: number);
    get uvX(): number;
    set uvX(value: number);
    get uvY(): number;
    set uvY(value: number);
    get colorR(): number;
    set colorR(value: number);
    get colorG(): number;
    set colorG(value: number);
    get colorB(): number;
    set colorB(value: number);
    trueIndex: number;
    _index: number;
    data: {
        [index: number]: number;
    };
    get index(): number;
    set index(index: number);
    constructor(data?: Float32Array);
    toJSON(): {
        position: number[];
        normal: number[];
        voxelData: number[];
        textureIndex: number[];
        uv: number[];
        color: number[];
        metadata: number[];
    };
}